Check for mod permissions, separate area for adding songs
Adding to/editing the songlist now checks for the moderator permission.
Added separate form for adding new items to the list since the built-in "add" functionality does not include grouped columns so they remain empty and cannot be saved because of missing values.
Added "created" date for songlist items so that a "recently added" section is possible.