Users: Some small updates to users to fix some issues.
Added way to add multiple users in a single SQL query, mainly used for
adding all of the chatlist in a single SQL cal instead of one for each
user.

Also stopped throwing an error if there is no access tokens. This
happens if a user who isn't the broadcaster logs in before the
broadcaster has logged in to setup auth tokens. Now just write to the
log and return undefined instead.