Added card trading event (!offer and !acceptoffer)
Also added "deleted" flag for cards on the stack so that a removal of a card does not reset the number of cards redeemed per week.

Added "add user from chat" for duel event since target user might not exist yet. While it doesn't make sense to duel such a user (because he will have 0 points), a "user does not exist" message would be misleading.