Added "base card name" property
By adding a card that has the "base card name" of another existing card, it's possible adding variants of cards, like extra rare "holographic" cards. Those extra cards still need to have a unique name (maybe prefixed with ⭐ or a similar symbol) to allow distinction when trading etc. but are supposed to be displayed instead of the more common cards when showing the list of cards to a user.