Fixed: User cookie is not consistently checked for undefined
When navigating between areas after the user cookie has expired (only stays active for a couple of minutes), the UI potentially crashes when accessing an undefined user cookie.

Also set broadcaster_user to 0 instead of clearing because clearing a cookie that does not exist results in a warning.

Made logging to file optional

Extended nodemon ignorelist to exclude chewiebot.db and log files