Added transaction for adding seasons
Transaction is needed to make sure that the entire process is either completed properly  or not at all, we don't want any half done things here. Also this speeds up the whole process quite a bit.
Note: This is not ideal, transaction should not be set globally with useTransaction() but instead passed as a parameter. Not sure how to properly do this without rewriting all repository classes.