Fixed calculating stream duration for a given time period
Start and end should both be at 0:00 so 7 full days are checked (adding the current time to start might miss a stream if the stream last week started earlier).
Also if stream runs over midnight (server time), the event list will be missing a "stream end" and not include the stream in the calculation.