Refactored user authorizations to be in separate table
For most users, the authorizations are typically empty so it's not necessary to have all information in the users table.
Also, adding further authorizations always requires a database change, which can be avoided.
The new table also saves the scope used for authorization, so it can be checked if an updated auth is needed when the required scopes have changed.
Also the streamlabs socket token was never encrypted which is fixed with the change.